# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Tạo Ảnh Đẹp API",
    "contact": {
      "email": "kaysilminer@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://taoanhdep.kaysil.dev"
    }
  ],
  "tags": [
    {
      "name": "wibu",
      "description": "Tạo ảnh avatar anime chữ ký phong cách anime"
    }
  ],
  "paths": {
    "/v1/wibu/create": {
      "get": {
        "tags": ["wibu"],
        "summary": "Tạo ảnh",
        "operationId": "wibuCreate",
        "parameters": [
          {
            "name": "id_nhanvat",
            "in": "query",
            "description": "ID của nhân vật",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 450
            }
          },
          {
            "name": "chu_ky",
            "in": "query",
            "description": "Chữ ký ở đầu ảnh",
            "required": true,
            "style": "form",
            "explode": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chu_nen",
            "in": "query",
            "description": "Chữ nền đằng sau nhân vật anime",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mau_nen",
            "in": "query",
            "description": "màu nền (tên hoặc mã hex)",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "format": "hex"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trả về ảnh khi tạo thành công",
            "content": {
              "image/png": {}
            }
          },
          "400": {
            "description": "Nhập sai/thiếu query parameters"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        }
      }
    },
    "/v1/wibu/list": {
      "get": {
        "tags": ["wibu"],
        "summary": "Lấy danh sách nhân vật",
        "operationId": "wibuList",
        "responses": {
          "200": {
            "description": "Lấy danh sách thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WibuListResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "WibuListResponse": {
        "required": ["id", "manufacturer", "name", "releaseDate"],
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "format": "number",
            "example": 200
          },
          "message": {
            "type": "string",
            "format": "string",
            "example": "Successful"
          },
          "data": {
            "type": "array",
            "format": "array",
            "items": {
              "$ref": "#/components/schemas/WibuListData"
            }
          }
        }
      },
      "WibuListData": {
        "required": ["name"],
        "type": "object",
        "properties": {
          "characterId": {
            "type": "number",
            "example": 0
          },
          "characterName": {
            "type": "string",
            "example": "hy tomboy"
          },
          "imageUrl": {
            "type": "string",
            "format": "url",
            "example": "https://kaysil.dev/hy-tomboy.png"
          }
        }
      }
    }
  }
}